fixed bug: cache update was not running on create/update location
moved scheduler for cache-update to LocationProxyRule.java
modified scheduler for cache-update to not run in TEST mode
